pyproject.toml
[tool.poetry]
name = "thepost"
version = "0.1.0"
description = "News site base system"
authors = ["Diogo"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
environs = {extras = ["django"], version = "^9.3.4"}
libsass = "^0.21.0"
django-compressor = "^2.4.1"
django-sass-processor = "^1.1"
gunicorn = "^20.1.0"
django-taggit = "^1.5.1"
Pillow = "^8.4.0"
django-ckeditor = "^6.1.0"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
